break PackingInput and Viewer components out of App.tsx
