+++ src/App.css
@@ -42,39 +42,6 @@
   margin-top: 5%;
 }
 
-.recipeBox {
-  width: 45%;
-  float: left;
-}
-
-.recipeJSON {
-  max-height: 300px;
-  width: 450px;
-  overflow-y: auto;
-}
-
-.configBox {
-  width: 45%;
-  float: right;
-}
-
-.configJSON {
-  max-height: 300px;
-  width: 450px;
-  overflow-y: auto;
-}
-
-textarea {
-  width: 450px;
-  height: 300px;
-  padding: 12px 20px;
-  box-sizing: border-box;
-  border: 2px solid #ccc;
-  border-radius: 4px;
-  font-size: 13px;
-  resize: none;
-}
-
 .logBox {
   display: flex;
   justify-content: space-between;
