+++ src/components/PackingInput/style.css
@@ -1,0 +1,32 @@
+.recipeBox {
+  width: 45%;
+  float: left;
+}
+
+.configBox {
+  width: 45%;
+  float: right;
+}
+
+.recipeJSON {
+  max-height: 300px;
+  width: 450px;
+  overflow-y: auto;
+}
+
+.configJSON {
+  max-height: 300px;
+  width: 450px;
+  overflow-y: auto;
+}
+
+textarea {
+  width: 450px;
+  height: 300px;
+  padding: 12px 20px;
+  box-sizing: border-box;
+  border: 2px solid #ccc;
+  border-radius: 4px;
+  font-size: 13px;
+  resize: none;
+}
