+++ src/components/Viewer/style.css
@@ -1,0 +1,5 @@
+.simularium-embed {
+    width: 1000px;
+    height: 600px;
+    border: 1px solid black;
+}
